st_setclientid.py: Lower down clients to 1023
This fixes error for all nfs4.0 tests:
nfs4lib.BadCompoundRes: operation OP_SETCLIENTID should return NFS4_OK, instead got NFS4ERR_DELAY

Neil Brown:
That is probably the "courteous server" stuff that landed in 6.0.
A consequence is that there is a limit of 1024 clients per GB of memory.
The test that is failing tries to create 1024 client and presumably his
the limit.

I get the same error with a 1G server VM, and not error with 3G.

Suggested-by: NeilBrown <neilb@suse.de>
Signed-off-by: Petr Vorel <pvorel@suse.cz>